meta-qti-eap: Fixes for systemd messgaes.
Journal was disabled for better KPI and the errors generated by the
service files trying to connect to journal socket for logging resolved
through this patch.

The groups render and kvm fixes are applicable for VT similar to
SA415m through this patch.

Srv is a serve folder. It holds the site's specific data, which is not required.

Change-Id: I30a4582c06d8e66d7a277adbe0776b3f912e4cb9