meta-qti-eap: Change user and group for tel.log files
As VT uses .ext4 file changing owners in recipe will be
overwritten by default while creating rootfs(/) and
userfs(/data) images. This conf is used to have our changes
reflected in userfs.

In the conf file, the user and group permissions of the tel.log
and tel.log.backup files are changed from root to system. Also
the file permissions modified to 0660.

Change-Id: I28056c2b842982694d2768d43ba5fac19b6ac12b